movie_histogram/requirements.txt
Django==2.0.7
numpy
django-model-utils